requirements.txt
django==5.0.1
psycopg2==2.9.9
django-environ==0.11.2
django-modeltranslation==0.18.11
pyTelegramBotAPI==4.15.3
django-import-export==3.3.7